4. Django Advanced Tutorial/requirements.txt
django==4.1.1
pillow==9.2.0
djangorestframework==3.14.0